Reduce warnings (#615)
* shared reduced warnings

* Reduce compile warnings

* update GHA scala version

* Upgrade scoverage for scala 2.13.10 support

* change wrong sbt-scoverage version

* format code

* Revert schema creation in bq test

* Fix compilation issue

* Fix jmh project setup

+++ guava/src/test/scala/magnolify/guava/test/FunnelDerivationSuite.scala
@@ -105,7 +105,7 @@ class BytesSink extends PrimitiveSink {
   }
 
   override def putByte(b: Byte): PrimitiveSink = {
-    oos.writeByte(b)
+    oos.writeByte(b.toInt)
     this
   }
 
@@ -124,7 +124,7 @@ class BytesSink extends PrimitiveSink {
   }
 
   override def putShort(s: Short): PrimitiveSink = {
-    oos.writeShort(s)
+    oos.writeShort(s.toInt)
     this
   }
 
@@ -154,7 +154,7 @@ class BytesSink extends PrimitiveSink {
   }
 
   override def putChar(c: Char): PrimitiveSink = {
-    oos.writeChar(c)
+    oos.writeChar(c.toInt)
     this
   }
